+++ src/main/java/sin/java/interfaces/MovementImplOldStyle.java
@@ -1,9 +1,6 @@
 package sin.java.interfaces;
 
-public class MovementImplOldStyle implements MovementInt{
-    /*
-    * Override annotation is used to implement the method how(), which can be avoided in default method
-    * */
+public class MovementImplOldStyle implements MovementOldStyle{
     @Override
     public String howOldStyle() {
         return "I am moving from Implementation Class - Old Style";
