+++ src/main/java/sin/java/interfaces/MovementOldStyle.java
@@ -1,0 +1,5 @@
+package sin.java.interfaces;
+
+public interface MovementOldStyle {
+    String howOldStyle();
+}
