--- src/main/java/sin/java/readpropertyfile/Main.java
@@ -1,8 +1,0 @@
-package sin.java.readpropertyfile;
-
-public class Main {
-    public static void main(String[] args) {
-        ReadPropertyFile readPropertyFile = new ReadPropertyFile();
-        System.out.println(readPropertyFile.readMyProperty("my-url"));
-    }
-}
