+++ src/test/java/sin/java/interfaces/MovementImplOldStyleTest.java
@@ -12,10 +12,4 @@ public class MovementImplOldStyleTest {
         MovementImplOldStyle movementImplOldStyle = new MovementImplOldStyle();
         assertThat(movementImplOldStyle.howOldStyle()).isEqualTo("I am moving from Implementation Class - Old Style");
     }
-
-    @Test
-    public void test_howDefaultNewStyle() {
-        MovementImplOldStyle movementImplOldStyle = new MovementImplOldStyle();
-        assertThat(movementImplOldStyle.how()).isEqualTo("I am implemented in Interface with default - New Style");
-    }
 }
